refactor(harness): extraer workflow-executor, run-lifecycle y run-helpers — fase 2 (04/T05)
Completa la modularización de run-engine.cjs (1.950 → 1.310 líneas, API
pública intacta): workflow-executor.cjs (ejecutor DAG: executeWorkflowRun
con retry por nodo, resolveWorkflowAgent, SYSTEM_AGENTS, nodos estáticos;
cableado por init({database, loadManyAgents})), run-lifecycle.cjs
(activeRunContexts, releaseRunContext, abortRun y cleanup de shutdown) y
run-helpers.cjs (helpers puros: detección de abort, parseo de tool args,
merge de usage LLM, patches de tool-steps — 6 tests unitarios en
test:security). Diagrama actualizado en docs/architecture.
